cleanup: move shared definitions into a dedicated header
